add edit entities
