+++ src/main/java/group/ship/blackshipstore/entity/ItemValues.java
@@ -1,0 +1,46 @@
+package group.ship.blackshipstore.entity;
+
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "item_values")
+public class ItemValues {
+
+    @EmbeddedId
+    @Column(name = "id")
+    private ItemsValuesArticlesKey id;
+
+    @ManyToOne(cascade = CascadeType.ALL)
+    @MapsId("itemsId")
+    @JoinColumn(name = "item_id")
+    private Item item;
+
+    @ManyToOne(cascade = CascadeType.ALL)
+    @MapsId("valueId")
+    @JoinColumn(name = "value_id")
+    private Value value;
+
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "article_id",referencedColumnName = "id")
+    private Article article;
+
+
+    public ItemValues() {
+    }
+
+    public Item getItem() {
+        return item;
+    }
+
+    public void setItem(Item item) {
+        this.item = item;
+    }
+
+    public Value getValue() {
+        return value;
+    }
+
+    public void setValue(Value value) {
+        this.value = value;
+    }
+}
