+++ src/main/java/group/ship/blackshipstore/repositories/ArticlesRepository.java
@@ -1,6 +1,7 @@
 package group.ship.blackshipstore.repositories;
 
 import group.ship.blackshipstore.entity.Article;
+import group.ship.blackshipstore.entity.ItemValues;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
